<Android App>找房页面逻辑（ongoing） 可以正确显示tag， 价格，图片等

+++ App_Android/app/src/main/java/com/kjs/skywalk/app_android/ClassDefine.java
@@ -70,8 +70,8 @@ public class ClassDefine {
         public int Livingrooms;
         public int Bathrooms;
         public double Acreage;
-        public int Rental;
-        public int Pricing;
+        public double Rental;
+        public double Pricing;
         public int CoverImage;
         public String CoverImageUrlS;
         public String CoverImageUrlM;
